refactor(web): split oversized analysis tables to satisfy 400-line gate
Extract helpers so ConditionDecisionsTable.tsx and PivotAnalysisTable.tsx
stay under the 400-line file-size limit:

- modelHeaderLabels.ts — model family inference, label formatting,
  buildModelHeaders, groupModelHeadersByFamily, hasGroupedFamilyVariants
- ConditionDecisionsTableHead.tsx — extracted <thead> subcomponent
- PivotAnalysisLegend.tsx — extracted legend subcomponent + LegendCounts

No behavior change. 1246 web tests pass.
